requirements.txt
requests
discord.py[voice]
pyqt5
sounddevice
SoundFile
pywin32